refactor app structure, hardcode local db json
