+++ .gitignore
@@ -6,3 +6,4 @@ __pycache__*
 htmlcov/
 
 *.env
+db/local.json
